working on fixing OCI conformance
